# api/openapi.yaml
openapi: 3.0.3
info:
  title: API ImmoConnect
  description: >
    API RESTful de la plateforme ImmoConnect (agence immobilière) :
    consultation des biens, favoris, messagerie, prise de rendez-vous de visite
    avec créneaux premium payants (Stripe), blog et volet Open Data.
    Documentation générée avec springdoc-openapi et servie par Swagger UI.
  version: 1.0.0
  contact:
    name: ImmoConnect — support API
    email: api@immoconnect.be
  license:
    name: Données Open Data sous licence CC BY 4.0
    url: https://creativecommons.org/licenses/by/4.0/deed.fr
servers:
  - url: https://www.immoconnect.be/api/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Développement local
...
security:
  - bearerAuth: []
paths:
...
  /biens:
    get:
      tags: [Biens]
      summary: Rechercher les biens (multicritères, paginé)
      security: []
      parameters:
        - { name: typeOffre, in: query, schema: { $ref: '#/components/schemas/TypeOffre' }, description: Sans ce critère, ventes et locations sont renvoyées ensemble }
        - { name: ville, in: query, schema: { type: string }, example: Ixelles, description: Début du nom de la commune ou de son code postal }
        - { name: categorieId, in: query, schema: { type: integer } }
        - { name: prixMin, in: query, schema: { type: number } }
        - { name: prixMax, in: query, schema: { type: number }, example: 350000, description: Prix de vente ou loyer mensuel maximum }
        - { name: chambresMin, in: query, schema: { type: integer } }
        - { name: superficieMin, in: query, schema: { type: number } }
        - { name: statut, in: query, schema: { type: string, enum: [disponible, sous_option, vendu, loue], default: disponible }, description: Un bien hors ligne (archive) n'est jamais listé }
        - { name: page, in: query, schema: { type: integer, default: 0 } }
        - { name: taille, in: query, schema: { type: integer, default: 20, maximum: 100 } }
        - { name: tri, in: query, schema: { type: string, example: 'prix,asc' } }
      responses:
        '200':
          description: Page de résultats
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageBiens' }
    post:
      tags: [Biens]
      summary: Créer une annonce (agent)
      description: >
        L'annonce est créée hors ligne (statut « archive ») : elle ne se publie qu'une fois sa
        première photo ajoutée (règle RA6). L'agent responsable est l'agent connecté.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BienCreation' }
      responses:
        '201':
          description: Annonce créée
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BienGestion' }
        '403': { $ref: '#/components/responses/ErreurDroits' }
        '422': { $ref: '#/components/responses/ErreurValidation' }
  /biens/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
    get:
      tags: [Biens]
      summary: Détail d'un bien
      security: []
      responses:
        '200':
          description: Le bien
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Bien' }
        '404': { $ref: '#/components/responses/ErreurIntrouvable' }
    put:
      tags: [Biens]
      summary: Modifier une annonce et son statut (agent responsable)
      description: >
        Le champ « statut » est facultatif. Publier exige au moins une photo (RA6) ; vendu, loué ou
        archivé est refusé tant que des visites sont prévues. L'annonce d'un autre agent renvoie 403.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BienCreation' }
      responses:
        '200':
          description: Annonce mise à jour
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BienGestion' }
        '403': { $ref: '#/components/responses/ErreurDroits' }
        '409': { $ref: '#/components/responses/ErreurConflit' }
        '422': { $ref: '#/components/responses/ErreurValidation' }
    delete:
      tags: [Biens]
      summary: Archiver une annonce (agent responsable)
      description: Suppression logique — l'annonce quitte la recherche et reste dans l'historique du back-office.
      responses:
        '204': { description: Annonce archivée }
        '403': { $ref: '#/components/responses/ErreurDroits' }
        '409': { $ref: '#/components/responses/ErreurConflit' }
...
  /agents/moi/biens/{id}:
    get:
      tags: [Biens]
      summary: Une de mes annonces, adresse exacte comprise (agent responsable)
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: L'annonce
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BienGestion' }
        '403': { $ref: '#/components/responses/ErreurDroits' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Jeton JWT obtenu via POST /auth/login (en-tête Authorization: Bearer <jeton>)"
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Clé API délivrée par l'administrateur (table cle_api) — révocable (règle RA12)
  responses:
    ErreurValidation:
      description: Données invalides (règles du dictionnaire de données)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Erreur' }
...
    ErreurDroits:
      description: Rôle insuffisant pour cette opération
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Erreur' }
...
    ErreurIntrouvable:
      description: Ressource inexistante
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Erreur' }
    ErreurConflit:
      description: Conflit d'état (créneau déjà réservé — A2, transition interdite — RA1)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Erreur' }
...
  schemas:
    StatutBien:
      type: string
      enum: [disponible, sous_option, vendu, loue, archive]
      description: « vendu » ne concerne qu'une vente, « loue » qu'une location
    TypeOffre:
      type: string
      enum: [vente, location]
      description: Nature de l'offre ; elle donne son sens au prix
    Peb:
      type: string
      enum: ['A++', 'A+', A, B, C, D, E, F, G]
      description: Classe énergétique du certificat PEB, obligatoire dans toute publicité immobilière
...
    Photo:
      type: object
      properties:
        id: { type: integer }
        url: { type: string, example: /storage/biens/42/photo-1.jpg }
        ordre: { type: integer, example: 1 }
        legende: { type: string, nullable: true, example: Séjour }
    Bien:
      type: object
      properties:
        id: { type: integer, example: 42 }
        titre: { type: string, example: Appartement 3 chambres — Ixelles }
        description: { type: string }
        categorie: { type: string, example: Appartement }
        typeOffre: { $ref: '#/components/schemas/TypeOffre' }
        prix: { type: number, example: 285000, description: Prix de vente, ou loyer mensuel d'une location }
        superficie: { type: number, example: 95 }
        nbChambres: { type: integer, example: 3 }
        peb: { $ref: '#/components/schemas/Peb' }
        adresse: { type: string, example: Rue du Bailli 27 }
        ville: { type: string, example: Ixelles }
        codePostal: { type: string, example: '1050' }
        latitude: { type: number, example: 50.8275 }
        longitude: { type: number, example: 4.3720 }
        statut: { $ref: '#/components/schemas/StatutBien' }
        publieLe: { type: string, format: date }
        agent:
          type: object
          properties:
            id: { type: integer, example: 101 }
            nomComplet: { type: string, example: Sarah Dubois }
        photos:
          type: array
          items: { $ref: '#/components/schemas/Photo' }
    BienCreation:
      type: object
      required: [categorieId, titre, description, prix, superficie, nbChambres, peb, adresse, ville, codePostal, latitude, longitude]
      properties:
        categorieId: { type: integer }
        typeOffre:
          allOf: [{ $ref: '#/components/schemas/TypeOffre' }]
          description: Facultatif — une nouvelle annonce est une vente, une annonce modifiée garde le sien
        titre: { type: string, maxLength: 150 }
        description: { type: string }
        prix: { type: number, minimum: 0, exclusiveMinimum: true, description: Prix de vente, ou loyer mensuel d'une location }
        superficie: { type: number, minimum: 0, exclusiveMinimum: true }
        nbChambres: { type: integer, minimum: 0, maximum: 20 }
        peb: { $ref: '#/components/schemas/Peb' }
        adresse: { type: string }
        ville: { type: string }
        codePostal: { type: string, pattern: '^[1-9][0-9]{3}$' }
        latitude: { type: number, minimum: -90, maximum: 90 }
        longitude: { type: number, minimum: -180, maximum: 180 }
    PageBiens:
      type: object
      properties:
        contenu:
          type: array
          items: { $ref: '#/components/schemas/Bien' }
        page: { type: integer, example: 0 }
        taille: { type: integer, example: 20 }
        totalElements: { type: integer, example: 128 }
        totalPages: { type: integer, example: 7 }
...
    BienGestion:
      type: object
      description: Annonce vue par son agent — adresse exacte comprise, jamais renvoyée au public
      properties:
        id: { type: integer, example: 42 }
        titre: { type: string }
        description: { type: string }
        typeOffre: { $ref: '#/components/schemas/TypeOffre' }
        prix: { type: number, example: 285000, description: Prix de vente, ou loyer mensuel d'une location }
        superficie: { type: number, example: 95 }
        nbChambres: { type: integer, example: 3 }
        peb: { $ref: '#/components/schemas/Peb' }
        adresse: { type: string, example: Rue du Bailli 27 }
        ville: { type: string, example: Ixelles }
        codePostal: { type: string, example: '1050' }
        latitude: { type: number, example: 50.8275 }
        longitude: { type: number, example: 4.3720 }
        statut: { $ref: '#/components/schemas/StatutBien' }
        publieLe: { type: string, format: date }
        categorieId: { type: integer }
        categorie: { type: string, example: Appartement }
        photos: { type: array, items: { $ref: '#/components/schemas/Photo' } }
        indicateurs:
          type: object
          properties:
            vues: { type: integer, example: 143, description: Consultations de la fiche publique par les visiteurs et les membres }
            favoris: { type: integer, example: 6 }
            demandesEnAttente: { type: integer, example: 1 }
            visitesAVenir: { type: integer, example: 2 }
...
    Erreur:
      type: object
      description: Format RFC 7807 (application/problem+json)
      properties:
        type: { type: string, example: https://www.immoconnect.be/erreurs/creneau-indisponible }
        title: { type: string, example: Créneau indisponible }
        status: { type: integer, example: 409 }
        detail: { type: string, example: Ce créneau vient d'être réservé par un autre membre. }
        champs:
          type: object
          additionalProperties: { type: string }
          example: { email: L'adresse e-mail est invalide }